refactor(server/events): extract file-sync-registry.ts, fix spy target
Move fileSyncContexts Map, safeCleanup, setFileSyncContext, clearFileSyncContext
to file-sync-registry.ts. Add reattachFileSyncObserver (used by reattachObservers)
and resetForTesting. queue.ts re-exports set/clearFileSyncContext for backward
compat and delegates registry reset via fileSyncResetForTesting.

Apply pre-decided spy-target switch in file-opener-lifecycle.test.ts: spy now
targets fileSyncRegistryModule.setFileSyncContext rather than the re-export
in queueModule (Vitest ESM live bindings require the originating module).
Removed unused queueModule import (no other references in that file).
