+++ src/server/events/queue.ts
@@ -22,21 +22,21 @@ import {
   Y_MAP_USER_AWARENESS,
 } from "../../shared/constants.js";
 import type { Annotation, AnnotationReply, ChatMessage, FlatOffset } from "../../shared/types.js";
-import type { DocStore } from "../annotations/store.js";
-import {
-  type ObserverCleanupPhase,
-  registerAnnotationObserver,
-  type SyncContext,
-} from "../annotations/sync.js";
 import { sanitizeAnnotation } from "../mcp/annotations.js";
 import { getOpenDocs } from "../mcp/document-service.js";
 import { validateRange } from "../positions.js";
 import { getOrCreateDocument } from "../yjs/provider.js";
+import {
+  clearFileSyncContext,
+  resetForTesting as fileSyncResetForTesting,
+  reattachFileSyncObserver,
+  setFileSyncContext,
+} from "./file-sync-registry.js";
 import { FILE_SYNC_ORIGIN, MCP_ORIGIN } from "./origins.js";
 import type { TandemEvent } from "./types.js";
 import { generateEventId } from "./types.js";
 
-export { FILE_SYNC_ORIGIN, MCP_ORIGIN };
+export { clearFileSyncContext, FILE_SYNC_ORIGIN, MCP_ORIGIN, setFileSyncContext };
 
 /**
  * Read the user's configured selection dwell time from CTRL_ROOM.
@@ -331,93 +331,7 @@ export function reattachObservers(docName: string, newDoc: Y.Doc): void {
   // Hocuspocus doc swap. The previous observer was attached to the old
   // Y.Doc and is no longer reachable (Hocuspocus destroyed the old doc
   // in onLoadDocument); the cleanup we stashed would be a no-op anyway.
-  const oldCtx = fileSyncContexts.get(docName);
-  if (oldCtx) {
-    safeCleanup(docName, oldCtx.cleanup, "swap", "reattach");
-    const newCtx: SyncContext = {
-      ydoc: newDoc,
-      store: oldCtx.ctx.store,
-      docHash: oldCtx.ctx.docHash,
-      meta: oldCtx.ctx.meta,
-    };
-    const cleanup = registerAnnotationObserver(newCtx);
-    fileSyncContexts.set(docName, { ctx: newCtx, cleanup });
-  }
-}
-
-// --- File-sync observer registry (durable annotations) ---
-
-/**
- * Track per-doc SyncContext so `reattachObservers` can re-register the
- * annotation file-writer observer against the new Y.Doc after a Hocuspocus
- * doc swap. Stored alongside the cleanup handle from the prior
- * `registerAnnotationObserver` call so we can dispose it deterministically.
- */
-const fileSyncContexts = new Map<
-  string,
-  { ctx: SyncContext; cleanup: (phase?: ObserverCleanupPhase) => void }
->();
-
-/**
- * Run a file-sync observer cleanup in a try/catch with a uniform log line.
- * Cleanups can throw if the underlying Y.Doc was already destroyed (e.g. a
- * Hocuspocus reload beat us to it) — that's harmless, but we want the logs
- * to be consistent enough to grep.
- */
-function safeCleanup(
-  docName: string,
-  cleanup: (phase?: ObserverCleanupPhase) => void,
-  phase: ObserverCleanupPhase,
-  logTag: string,
-): void {
-  try {
-    cleanup(phase);
-  } catch (err) {
-    console.warn("[EventQueue] file-sync cleanup threw during %s for %s:", logTag, docName, err);
-  }
-}
-
-/**
- * Register the durable-annotation sync context for a document. Called by the
- * file-opener after `loadAndMerge` returns its observer cleanup. The cleanup
- * passed here is what gets invoked on `clearFileSyncContext` or the next
- * `reattachObservers` call.
- *
- * Callers don't need to think about the swap-vs-close distinction: the queue
- * picks the phase at teardown time based on which entrypoint ran.
- */
-export function setFileSyncContext(
-  docName: string,
-  ctx: SyncContext,
-  cleanup: (phase?: ObserverCleanupPhase) => void,
-): void {
-  // Dispose any prior entry first so we never leak observers on duplicate
-  // registration (e.g., forceReload paths that re-run loadAndMerge). Normal
-  // flow pre-clears via `clearFileSyncContext`, so this branch is defensive;
-  // `"close"` is correct because we're replacing — not rebinding — the
-  // context, so the prior docHash's tombstones belong to a superseded state.
-  const existing = fileSyncContexts.get(docName);
-  if (existing) {
-    safeCleanup(docName, existing.cleanup, "close", "replace");
-  }
-  fileSyncContexts.set(docName, { ctx, cleanup });
-}
-
-/**
- * Drop the file-sync context for a document (on close or force-reload prep).
- * Returns the dropped `{ store, docHash }` so callers can flush/clear the
- * durable store without recomputing the hash or minting a transient handle.
- * Returns `undefined` if no context was registered (e.g. feature flag off,
- * or `wireAnnotationStore` failed during open).
- */
-export function clearFileSyncContext(
-  docName: string,
-): { store: DocStore; docHash: string } | undefined {
-  const entry = fileSyncContexts.get(docName);
-  if (!entry) return undefined;
-  safeCleanup(docName, entry.cleanup, "close", "clear");
-  fileSyncContexts.delete(docName);
-  return { store: entry.ctx.store, docHash: entry.ctx.docHash };
+  reattachFileSyncObserver(docName, newDoc);
 }
 
 // --- CTRL_ROOM observers (chat + document meta) ---
@@ -573,22 +487,22 @@ export function reattachCtrlObservers(): void {
 
 /** Reset all module state. For tests only — do not call in production. */
 export function resetForTesting(): void {
+  // 1. Clear data-only collections (observer cleanups don't touch these)
   buffer.length = 0;
   subscribers.clear();
   emittedPayloadIds.clear();
   selectionBuffer.clear();
+
+  // 2. Run per-doc observer cleanups, then clear the map that holds them
   for (const cleanups of docObservers.values()) {
     for (const cleanup of cleanups) cleanup();
   }
   docObservers.clear();
+
+  // 3. Run CTRL cleanups, then reset the array that holds them
   for (const cleanup of ctrlCleanups) cleanup();
   ctrlCleanups = [];
-  for (const entry of fileSyncContexts.values()) {
-    try {
-      entry.cleanup("close");
-    } catch {
-      /* ignore — test cleanup should not throw on pre-torn-down docs */
-    }
-  }
-  fileSyncContexts.clear();
+
+  // 4. Delegate registry reset (CRITICAL — do not forget)
+  fileSyncResetForTesting();
 }
